different types can be input

+++ include/bit.cl
@@ -1,4 +1,4 @@
-__kernel void bitonic_sort(__global int *data, const int n) {
+__kernel void bitonic_sort(__global double *data, const int n) {
     int index = get_global_id(0);
 
     for (int k = 2; k <= n; k *= 2) {
@@ -17,7 +17,6 @@ __kernel void bitonic_sort(__global int *data, const int n) {
                         int temp = data[index];
                         data[index] = data[ixj];
                         data[ixj] = temp;
-                        
                     }
                 }
             }
